Fix background not staying behind the code when scrolling horizontally
This commit fixes a bug introduced in e2032e19b05 which changed the
previous table layout to a pure CSS-based layout. Because the new layout
used a "position: fixed" background, the background would not scroll
when code lines were more wide than the browser window. This has been
fixed by making the background "position: absolute" and moving it into
the container DIV. Because the container DIV does not expand if its
content is wider than the browser window though, I had to set the body
background color to the same color than the code background and the
container background to white, so that the left (comment) part stays
white.